+++ Sources/KeyboardKit/System/SystemKeyboardActionButtonContent.swift
@@ -54,7 +54,7 @@ public struct SystemKeyboardActionButtonContent: View {
         } else if action == .space {
             spaceView
         } else if let image = appearance.buttonImage(for: action) {
-            image
+            image.scaleEffect(appearance.buttonImageScaleFactor(for: action))
         } else if let text = appearance.buttonText(for: action) {
             textView(for: text)
         } else {