+++ RELEASE_NOTES.md
@@ -40,6 +40,8 @@ Most of the changes only affect functionality that is mostly used internally, bu
 * `DisabledCalloutActionProvider` is now `public`.
 * `InterfaceOrientation` is a new multi-platform version of `UIInterfaceOrientation`.
 * `KeyboardAppearance` has a new `keyboardEdgeInsets` property.
+* `KeyboardAction` has a new `isAlphabeticKeyboardTypeAction` property.
+* `KeyboardAction` has a new `isKeyboardTypeAction` function.
 * `KeyboardContext` has a new `screenSize` parameter.
 * `KeyboardContext` has a new `interfaceOrientation` parameter.
 * `KeyboardTextView` has a new `hasFocus` binding.
@@ -61,7 +63,8 @@ Most of the changes only affect functionality that is mostly used internally, bu
 * `KeyboardAction.primary` now applies autocomplete suggestions if it's a system action.
 * `KeyboardContext` controller-based initializer is marked as a convenience initializer.
 * `SystemKeyboard` no longer depends on `AnyView`.
-* `StandardKeyboardAppearance` will apply a correct bottom padding to the system keyboard.
+* `StandardKeyboardAppearance` now applies insets to system keyboards.
+* `StandardKeyboardAppearance` now applies larger font sizes to some system keyboards keys on iPad devices.
 * `SystemKeyboardSpaceButtonContent` no longer depends on `AnyView`.  
 
 ### 🗑 Deprecations